changes on character stats
