--- .plastic/plastic.branchexplorer
@@ -1,26 +1,0 @@
-[Default]
-display.changestats.visible=false
-display.navigator.visible=false
-display.options.branch_level_filter=-1
-display.options.changeset_color=1
-display.options.dag_mergelinks=false
-display.options.draw_branches=true
-display.options.draw_crossbranch_changeset_links=true
-display.options.draw_labels=true
-display.options.draw_merge_links=true
-display.options.draw_only_relevant_changesets=false
-display.options.draw_taskinfo=false
-display.options.end_date=638375039990000000
-display.options.end_date_enablement=false
-display.options.full_branch_names=true
-display.options.layout_mode=0
-display.options.show_excluded_replication_sources=false
-display.options.show_parent_to_child_arrows=false
-display.options.start_date=638348256000000000
-display.options.timesince=AGivenDate
-display.optionspanel.visible=false
-display.optionspanel.visibletabindex=0
-display.properties.legend=false
-display.properties.zoom=0.9
-filters.conditional.numrules=0
-
