+++ Assets/Scripts/Attributes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f5324245c9529e44ea638c2954b23698
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
