+++ Assets/Scripts/Attributes/Attribute.cs
@@ -1,0 +1,13 @@
+using UnityEngine;
+#if UNITY_EDITOR
+using UnityEditor;
+#endif
+
+public class Attribute : ScriptableObject
+{
+
+#if UNITY_EDITOR
+    public virtual void DoLayout() { }
+#endif
+
+}
