+++ Assets/Scripts/Attributes/Attribute.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 484469d0c24740c408ee7fbf9595242e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
