+++ Assets/Scripts/Character/Dozer.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8436a30d6aad4ee4d92e603f44f7fed6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
