+++ Assets/Scripts/Character/Player.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f9ed64a9595247b48829b5d186083d74
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
